# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/MessageEndpoint.java

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/SimpleHashtagService.java
@@ -3,8 +3,8 @@ package at.ac.tuwien.sepm.groupphase.backend.service.impl;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Hashtag;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Message;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Spot;
+import at.ac.tuwien.sepm.groupphase.backend.exception.ValidationException;
 import at.ac.tuwien.sepm.groupphase.backend.repository.HashtagRepository;
-import at.ac.tuwien.sepm.groupphase.backend.repository.MessageRepository;
 import at.ac.tuwien.sepm.groupphase.backend.service.HashtagService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,32 +24,29 @@ public class SimpleHashtagService implements HashtagService {
     private final String hashtagPattern = "(#[A-Za-z0-9]+)";
 
     @Override
-    public Hashtag getByName(String name) {
-        return hashtagRepository.getHashtagByNameEquals(name);
+    public Hashtag getByName(String name) throws ValidationException {
+        if (name == null || name.isEmpty() || name.isBlank()) {
+            throw new ValidationException("Hashtag must have a name");
+        }
+        Hashtag h =hashtagRepository.getHashtagByNameEquals(name);
+    return h;
     }
 
     @Override
-    public Hashtag getOneById(Long id) {
-        return hashtagRepository.getOneById(id);
-    }
-
-    @Override
-    public Hashtag create(Hashtag hashtag) {
-        return hashtagRepository.save(hashtag);
-    }
-
-    @Override
-    public void acquireHashtags(Message message){
+    public void acquireHashtags(Message message) throws ValidationException {
+        if (message == null || message.getContent() == null || message.getContent().isBlank() || message.getContent().isEmpty()) {
+            throw new ValidationException("the message must have content");
+        }
         String[] words = message.getContent().split("\\s+");
         List<String> hashtags = new ArrayList<>();
-        for(String word : words){
-            if (Pattern.matches(hashtagPattern, word)){
+        for (String word : words) {
+            if (Pattern.matches(hashtagPattern, word)) {
                 hashtags.add(word.substring(1));
             }
         }
 
-        for(String hashtag : hashtags){
-            if (hashtagRepository.findHashtagByName(hashtag).isPresent()){
+        for (String hashtag : hashtags) {
+            if (hashtagRepository.findHashtagByName(hashtag).isPresent()) {
                 Hashtag hashtag1 = hashtagRepository.findHashtagByName(hashtag).get();
                 hashtag1.addMessage(message);
                 hashtagRepository.save(hashtag1);
@@ -62,36 +59,45 @@ public class SimpleHashtagService implements HashtagService {
     }
 
     @Override
-    public void acquireHashtags(Spot spot) {
-        if(spot.getDescription()!=null) {
-            String[] words = spot.getDescription().split("\\s+");
-            List<String> hashtags = new ArrayList<>();
-            for (String word : words) {
-                if (Pattern.matches(hashtagPattern, word)) {
-                    hashtags.add(word.substring(1));
+    public void acquireHashtags(Spot spot) throws ValidationException {
+        if (spot != null) {
+            if (spot.getDescription() != null && !spot.getDescription().isBlank() && !spot.getDescription().isEmpty()) {
+                String[] words = spot.getDescription().split("\\s+");
+                List<String> hashtags = new ArrayList<>();
+                for (String word : words) {
+                    if (Pattern.matches(hashtagPattern, word)) {
+                        hashtags.add(word.substring(1));
+                    }
                 }
-            }
 
-            for (String hashtag : hashtags) {
-                if (hashtagRepository.findHashtagByName(hashtag).isPresent()) {
-                    Hashtag hashtag1 = hashtagRepository.findHashtagByName(hashtag).get();
-                    hashtag1.addSpot(spot);
-                    hashtagRepository.save(hashtag1);
-                } else {
-                    Hashtag hashtag1 = new Hashtag(hashtag);
-                    hashtag1.addSpot(spot);
-                    hashtagRepository.save(hashtag1);
+                for (String hashtag : hashtags) {
+                    if (hashtagRepository.findHashtagByName(hashtag).isPresent()) {
+                        Hashtag hashtag1 = hashtagRepository.findHashtagByName(hashtag).get();
+                        hashtag1.addSpot(spot);
+                        hashtagRepository.save(hashtag1);
+                    } else {
+                        Hashtag hashtag1 = new Hashtag(hashtag);
+                        hashtag1.addSpot(spot);
+                        hashtagRepository.save(hashtag1);
+                    }
                 }
+            } else {
+                throw new ValidationException("Spot must have a description");
             }
+        } else {
+            throw new ValidationException("Spot must exist");
         }
     }
 
     @Override
-    public void deleteMessageInHashtags(Message message){
+    public void deleteMessageInHashtags(Message message) throws ValidationException {
+        if (message == null) {
+            throw new ValidationException("the message must not be null");
+        }
         List<Hashtag> hashtags = hashtagRepository.findHashtagsByMessagesListContains(message);
-        for (Hashtag hashtag : hashtags){
+        for (Hashtag hashtag : hashtags) {
             hashtag.deleteMessage(message.getId());
-            if (hashtag.getMessagesList().isEmpty() && hashtag.getSpotsList().isEmpty()){
+            if (hashtag.getMessagesList().isEmpty() && hashtag.getSpotsList().isEmpty()) {
                 hashtagRepository.delete(hashtag);
             } else {
                 hashtagRepository.save(hashtag);
@@ -100,11 +106,14 @@ public class SimpleHashtagService implements HashtagService {
     }
 
     @Override
-    public void deleteSpotInHashtags(Spot spot) {
+    public void deleteSpotInHashtags(Spot spot) throws ValidationException {
+        if (spot == null) {
+            throw new ValidationException("the spot must not be null");
+        }
         List<Hashtag> hashtags = hashtagRepository.findHashtagsBySpotsListContains(spot);
-        for (Hashtag hashtag : hashtags){
+        for (Hashtag hashtag : hashtags) {
             hashtag.deleteSpot(spot.getId());
-            if (hashtag.getMessagesList().isEmpty() && hashtag.getSpotsList().isEmpty()){
+            if (hashtag.getMessagesList().isEmpty() && hashtag.getSpotsList().isEmpty()) {
                 hashtagRepository.delete(hashtag);
             } else {
                 hashtagRepository.save(hashtag);
